pyproject.toml
[tool.poetry]
name = "tame"
version = "0.1.0"
description = ""
authors = [
    "marios1861 <44480600+marios1861@users.noreply.github.com>",
]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8"
matplotlib = ">=3.7.1"
numpy = ">=1.24.2"
opencv-python = ">=4.8.0.76"
Pillow = ">=9.5.0"
scikit-learn = ">=1.2.2"
torch = ">=2.0.0, !=2.0.1 !=2.1.0"
torchvision = ">=0.15.1"
tqdm = ">=4.65.0"
PyYAML = ">=6.0"
pandas = ">=2.0.0"
lightning = ">=2.1.2"
torchmetrics = ">=0.11.4"
tensorboard = ">=2.12.2"
timm = ">=0.6.13"
python-dotenv = ">=1.0.0"
torchshow = "^0.5.1"
wandb = ">=0.16.1"
grad-cam = {path = "pytorch-grad-cam"}
einops = ">=0.7.0"
scikit-image = "*"
softadapt = {git = "https://github.com/dr-aheydari/SoftAdapt.git"}

[tool.poetry.scripts]
tame = 'tame.scripts:main'
tame-auto = 'tame.train:cli_main'

[tool.poetry.group.dev.dependencies]
torch-tb-profiler = ">=0.4.1"
ipykernel = ">=6.25.0"
nvitop = "^1.3.1"
pre-commit = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
